- [x] tests passed
- [x] passes `git diff upstream/master -u -- "*.py" | flake8 --diff`

The internals of `to_datetime` is getting a bit unwieldy, so split out the `_convert_listlike` logic and origin shifting logic to ` _convert_listlike_datetime` and `adjust_to_origin` methods respectively outside of `to_datetime`. The logic was not changed.