@pep8speaks 2018-07-02
Hello @mroeschke! Thanks for submitting the PR.

 - In the file [`pandas/core/tools/datetimes.py`](https://github.com/pandas-dev/pandas/blob/3e0c36b56c217d3297e1e7cb6a75ae2d046672ab/pandas/core/tools/datetimes.py), following are the PEP8 issues :

> [Line 243:17](https://github.com/pandas-dev/pandas/blob/3e0c36b56c217d3297e1e7cb6a75ae2d046672ab/pandas/core/tools/datetimes.py#L243): [E722](https://duckduckgo.com/?q=pep8%20E722) do not use bare except'
> [Line 314:9](https://github.com/pandas-dev/pandas/blob/3e0c36b56c217d3297e1e7cb6a75ae2d046672ab/pandas/core/tools/datetimes.py#L314): [E722](https://duckduckgo.com/?q=pep8%20E722) do not use bare except'