@jreback 2018-07-02
lgtm. can you do a perf check to make sure nothing changed (as the caching logic was lightly touched here)