@mroeschke 2018-07-03
My asv setup is still a little broken, but here's a benchmark showing no performance hit (with cache)

```
In [1]: from pandas import *

In [3]: N = 100

In [4]: dup_string_with_tz = ['2000-02-11 15:00:00-0800'] * N

# this branch
In [5]: %timeit to_datetime(dup_string_with_tz, cache=True)
1.31 ms ± 44.5 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [6]: %timeit to_datetime(dup_string_with_tz, cache=False)
2.3 ms ± 41.2 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

# Master
In [4]: %timeit to_datetime(dup_string_with_tz, cache=True)
1.29 ms ± 13.3 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [5]: %timeit to_datetime(dup_string_with_tz, cache=False)
2.31 ms ± 68.9 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```