@jreback 2018-07-03
thanks @mroeschke happily take refactorings to clean up things / make more readable